move collect to a more appropriate package
